Updated Get Model

+++ eyeflow_sdk/edge_client.py
@@ -132,14 +132,10 @@ def get_flow(app_token, flow_id):
 # ---------------------------------------------------------------------------------------------------------------------------------
 
 
-def get_model(app_token, dataset_id, model_folder):
+def get_model(app_token, dataset_id, model_folder, model_type="tensorflow"):
     local_doc = None
     try:
         log.info(f"Get model {dataset_id}")
-
-        folder_path = Path(model_folder + '/' + dataset_id)
-        if not folder_path.is_dir():
-            folder_path.mkdir(parents=True, exist_ok=True)
 
         local_cache = os.path.join(model_folder, dataset_id + '.json')
         if os.path.isfile(local_cache):
@@ -147,6 +143,7 @@ def get_model(app_token, dataset_id, model_folder):
                 local_doc = json.load(fp)
 
         endpoint = jwt.decode(app_token, options={"verify_signature": False})['endpoint']
+
         url = f"{endpoint}/published-model/{dataset_id}/"
         msg_headers = {'Authorization' : f'Bearer {app_token}'}
         payload = {"download_url": False}
@@ -175,17 +172,38 @@ def get_model(app_token, dataset_id, model_folder):
 
         model_doc = response.json()
 
-        dest_filename = os.path.join(model_folder, dataset_id + ".tar.gz")
-        download_file(model_doc["download_url"], dest_filename)
+        if "model_list" in model_doc:
+            for model_data in model_doc["model_list"]:
+                if model_data["info"].get("type", "") == model_type:
+                    download_url = model_data["download_url"]
+                    dest_filename = os.path.join(model_folder, model_data["info"]["file"])
+                    break
+            else:
+                log.error(f"Did not find model type {model_type} in {dataset_id} document")
+                return None
+        elif "download_url" in model_doc:
+            download_url = model_doc["download_url"]
+            dest_filename = os.path.join(model_folder, dataset_id + ".tar.gz")
+        else:
+            log.error("Get model response dont have model_list or download_url keys")
+            return None
+
+        download_file(download_url, dest_filename)
 
         # expand_file
-        call([
-            "tar",
-            "-xzf", dest_filename,
-            "--directory", folder_path
-        ])
+        if (dest_filename.endswith('tar.gz')):
 
-        os.remove(dest_filename)
+            folder_path = Path(model_folder + '/' + dataset_id)
+            if not folder_path.is_dir():
+                folder_path.mkdir(parents=True, exist_ok=True)
+
+            call([
+                "tar",
+                "-xzf", dest_filename,
+                "--directory", folder_path
+            ])
+
+            os.remove(dest_filename)
 
         if os.path.isfile(local_cache):
             os.remove(local_cache)
@@ -293,7 +311,7 @@ def upload_model(
         os.remove(hist_filename)
 
         if response.status_code != 201:
-            raise Exception(f"Failing upload model. Request URL: {url}. Response Json: {response.json()}")
+            raise Exception(f"Failing upload model. Response Json: {response.json()}")
 
         return dataset_id
